Adding all missing settings

+++ Appium/Server/Model/AppiumModel.m
@@ -10,8 +10,8 @@
 
 #import <Foundation/Foundation.h>
 #import "AppiumAppDelegate.h"
-#import "Utility.h"
 #import "AppiumPreferencesFile.h"
+#import "Utility.h"
 
 #pragma  mark - Model
 
@@ -47,6 +47,18 @@ BOOL _isServerListening;
 
 -(BOOL) androidFullReset { return [_defaults boolForKey:APPIUM_PLIST_ANDROID_FULL_RESET]; }
 -(void) setAndroidFullReset:(BOOL)fastReset { [_defaults setBool:fastReset forKey:APPIUM_PLIST_ANDROID_FULL_RESET]; }
+
+-(NSString*) androidKeyAlias { return [_defaults stringForKey:APPIUM_PLIST_ANDROID_KEY_ALIAS]; }
+-(void) setAndroidKeyAlias:(NSString *)androidKeyAlias { [_defaults setValue:androidKeyAlias forKey:APPIUM_PLIST_ANDROID_KEY_ALIAS]; }
+
+-(NSString*) androidKeyPassword { return [_defaults stringForKey:APPIUM_PLIST_ANDROID_KEY_PASSWORD]; }
+-(void) setAndroidKeyPassword:(NSString *)androidKeyPassword { [_defaults setValue:androidKeyPassword forKey:APPIUM_PLIST_ANDROID_KEY_PASSWORD]; }
+
+-(NSString*) androidKeystorePassword { return [_defaults stringForKey:APPIUM_PLIST_ANDROID_KEYSTORE_PASSWORD]; }
+-(void) setAndroidKeystorePassword:(NSString *)androidKeystorePassword { [_defaults setValue:androidKeystorePassword forKey:APPIUM_PLIST_ANDROID_KEYSTORE_PASSWORD]; }
+
+-(NSString*) androidKeystorePath { return [_defaults stringForKey:APPIUM_PLIST_ANDROID_KEYSTORE_PATH]; }
+-(void) setAndroidKeystorePath:(NSString *)androidKeystorePath { [_defaults setValue:androidKeystorePath forKey:APPIUM_PLIST_ANDROID_KEYSTORE_PATH]; }
 
 -(NSString*) androidPackage { return [_defaults stringForKey:APPIUM_PLIST_ANDROID_PACKAGE]; }
 -(void) setAndroidPackage:(NSString *)androidPackage { [_defaults setValue:androidPackage forKey:APPIUM_PLIST_ANDROID_PACKAGE]; }
@@ -118,6 +130,15 @@ BOOL _isServerListening;
 -(BOOL) keepArtifacts { return [_defaults boolForKey:APPIUM_PLIST_KEEP_ARTIFACTS]; }
 -(void) setKeepArtifacts:(BOOL)keepArtifacts { [_defaults setBool:keepArtifacts forKey:APPIUM_PLIST_KEEP_ARTIFACTS];}
 
+-(BOOL) killProcessesUsingPort { return [_defaults boolForKey:APPIUM_PLIST_KILL_PROCESSES_USING_PORT]; }
+-(void) setKillProcessesUsingPort:(BOOL)killProcessesUsingPort { [_defaults setBool:killProcessesUsingPort forKey:APPIUM_PLIST_KILL_PROCESSES_USING_PORT];}
+
+-(NSString*) logFile { return [_defaults stringForKey:APPIUM_PLIST_LOG_FILE]; }
+-(void) setLogFile:(NSString *)logFile { [_defaults setValue:logFile forKey:APPIUM_PLIST_LOG_FILE]; }
+
+-(NSString*) logWebHook { return [_defaults stringForKey:APPIUM_PLIST_LOG_WEBHOOK]; }
+-(void) setLogWebHook:(NSString *)logWebHook { [_defaults setValue:logWebHook forKey:APPIUM_PLIST_LOG_WEBHOOK]; }
+
 -(NSNumber*) nodeDebugPort { return [NSNumber numberWithInt:[[_defaults stringForKey:APPIUM_PLIST_NODEJS_DEBUG_PORT] intValue]]; }
 -(void) setNodeDebugPort:(NSNumber *)nodeDebugPort { [[NSUserDefaults standardUserDefaults] setValue:nodeDebugPort forKey:APPIUM_PLIST_NODEJS_DEBUG_PORT]; }
 
@@ -125,6 +146,9 @@ BOOL _isServerListening;
 -(void) setOrientationToForce:(iOSOrientation)orientationToForce {[self setOrientationToForceString:(orientationToForce == iOSOrientation_Landscape ? APPIUM_PLIST_FORCE_ORIENTATION_LANDSCAPE : APPIUM_PLIST_FORCE_ORIENTATION_PORTRAIT)]; }
 -(NSString*) orientationToForceString { return [[_defaults valueForKey:APPIUM_PLIST_ORIENTATION] isEqualToString:APPIUM_PLIST_FORCE_ORIENTATION_LANDSCAPE] ? APPIUM_PLIST_FORCE_ORIENTATION_LANDSCAPE : APPIUM_PLIST_FORCE_ORIENTATION_PORTRAIT ; }
 -(void) setOrientationToForceString:(NSString *)orientationToForceString { [_defaults setValue:orientationToForceString forKey:APPIUM_PLIST_ORIENTATION]; }
+
+-(BOOL) overrideExistingSessions { return [_defaults boolForKey:APPIUM_PLIST_OVERRIDE_EXISTING_SESSIONS]; }
+-(void) setOverrideExistingSessions:(BOOL)overrideExistingSessions { [_defaults setBool:overrideExistingSessions forKey:APPIUM_PLIST_OVERRIDE_EXISTING_SESSIONS]; }
 
 -(Platform)platform
 {
@@ -144,10 +168,7 @@ BOOL _isServerListening;
 -(void) setPort:(NSNumber *)port { [[NSUserDefaults standardUserDefaults] setValue:port forKey:APPIUM_PLIST_SERVER_PORT]; }
 
 -(BOOL) prelaunchApp { return [_defaults boolForKey:APPIUM_PLIST_PRELAUNCH]; }
--(void) setPrelaunchApp:(BOOL)preLaunchApp
-{
-	[_defaults setBool:preLaunchApp forKey:APPIUM_PLIST_PRELAUNCH];
-}
+-(void) setPrelaunchApp:(BOOL)preLaunchApp { [_defaults setBool:preLaunchApp forKey:APPIUM_PLIST_PRELAUNCH]; }
 
 -(BOOL) resetApplicationState { return [_defaults boolForKey:APPIUM_PLIST_RESET_APPLICATION_STATE]; }
 -(void) setResetApplicationState:(BOOL)resetApplicationState { [_defaults setBool:resetApplicationState forKey:APPIUM_PLIST_RESET_APPLICATION_STATE]; }
@@ -161,6 +182,9 @@ BOOL _isServerListening;
 -(NSNumber*) selendroidPort { return [NSNumber numberWithInt:[[_defaults stringForKey:APPIUM_PLIST_SELENDROID_PORT] intValue]]; }
 -(void) setSelendroidPort:(NSNumber *)selendroidPort{ [[NSUserDefaults standardUserDefaults] setValue:selendroidPort forKey:APPIUM_PLIST_SELENDROID_PORT]; }
 
+-(NSString*) seleniumGridConfigFile { return [_defaults stringForKey:APPIUM_PLIST_SELENIUM_GRID_CONFIG_FILE]; }
+-(void) setSeleniumGridConfigFile:(NSString *)seleniumGridConfigFile { [_defaults setValue:seleniumGridConfigFile forKey:APPIUM_PLIST_SELENIUM_GRID_CONFIG_FILE]; }
+
 -(NSString*) udid {return [_defaults stringForKey:APPIUM_PLIST_UDID];}
 -(void) setUdid:(NSString *)udid { [ _defaults setValue:udid forKey:APPIUM_PLIST_UDID]; }
 
@@ -169,6 +193,9 @@ BOOL _isServerListening;
 
 -(BOOL) useAndroidDeviceReadyTimeout { return [_defaults boolForKey:APPIUM_PLIST_USE_ANDROID_DEVICE_READY_TIMEOUT]; }
 -(void) setUseAndroidDeviceReadyTimeout:(BOOL)useAndroidDeviceReadyTimeout { [_defaults setBool:useAndroidDeviceReadyTimeout forKey:APPIUM_PLIST_USE_ANDROID_DEVICE_READY_TIMEOUT]; }
+
+-(BOOL) useAndroidKeystore { return [_defaults boolForKey:APPIUM_PLIST_USE_ANDROID_KEYSTORE]; }
+-(void) setUseAndroidKeystore:(BOOL)useAndroidKeystore { [_defaults setBool:useAndroidKeystore forKey:APPIUM_PLIST_USE_ANDROID_KEYSTORE]; }
 
 -(BOOL) useAndroidPackage {	return [_defaults boolForKey:APPIUM_PLIST_USE_ANDROID_PACKAGE]; }
 -(void) setUseAndroidPackage:(BOOL)useAndroidPackage { [_defaults setBool:useAndroidPackage forKey:APPIUM_PLIST_USE_ANDROID_PACKAGE]; }
@@ -193,6 +220,12 @@ BOOL _isServerListening;
 
 -(BOOL) useExternalNodeJSBinary { return self.developerMode && [_defaults boolForKey:APPIUM_PLIST_USE_EXTERNAL_NODEJS_BINARY]; }
 -(void) setUseExternalNodeJSBinary:(BOOL)useCustomNodeJSBinary { [_defaults setBool:useCustomNodeJSBinary forKey:APPIUM_PLIST_USE_EXTERNAL_NODEJS_BINARY]; }
+
+-(BOOL) useLogFile { return [_defaults boolForKey:APPIUM_PLIST_USE_LOG_FILE]; }
+-(void) setUseLogFile:(BOOL)useLogFile { [_defaults setBool:useLogFile forKey:APPIUM_PLIST_USE_LOG_FILE]; }
+
+-(BOOL) useLogWebHook { return [_defaults boolForKey:APPIUM_PLIST_USE_LOG_WEBHOOK]; }
+-(void) setUseLogWebHook:(BOOL)useLogWebHook { [_defaults setBool:useLogWebHook forKey:APPIUM_PLIST_USE_LOG_WEBHOOK]; }
 
 -(BOOL) useMobileSafari { return [_defaults boolForKey:APPIUM_PLIST_USE_MOBILE_SAFARI]; }
 -(void) setUseMobileSafari:(BOOL)useMobileSafari { [_defaults setBool:useMobileSafari forKey:APPIUM_PLIST_USE_MOBILE_SAFARI]; }
@@ -223,6 +256,9 @@ BOOL _isServerListening;
 -(BOOL) useSelendroidPort { return [_defaults boolForKey:APPIUM_PLIST_USE_SELENDROID_PORT]; }
 -(void) setUseSelendroidPort:(BOOL)useSelendroidPort { [_defaults setBool:useSelendroidPort forKey:APPIUM_PLIST_USE_SELENDROID_PORT]; }
 
+-(BOOL) useSeleniumGridConfigFile { return [_defaults boolForKey:APPIUM_PLIST_USE_SELENIUM_GRID_CONFIG_FILE]; }
+-(void) setUseSeleniumGridConfigFile:(BOOL)useSeleniumGridConfigFile { [_defaults setBool:useSeleniumGridConfigFile forKey:APPIUM_PLIST_USE_SELENIUM_GRID_CONFIG_FILE]; }
+
 -(BOOL) useUDID { return [_defaults boolForKey:APPIUM_PLIST_USE_UDID]; }
 -(void) setUseUDID:(BOOL)useUDID { [_defaults setBool:useUDID forKey:APPIUM_PLIST_USE_UDID]; }
 
@@ -241,12 +277,16 @@ BOOL _isServerListening;
 
 -(BOOL)startServer
 {
-    NSString* script = [NSString stringWithFormat: @"kill `lsof -t -i:%@`", self.port];
-    system([script UTF8String]);
-    
     if ([self killServer])
     {
         return NO;
+    }
+    
+    // kill any processes using the appium server port
+    if (self.killProcessesUsingPort)
+    {
+        NSString* script = [NSString stringWithFormat: @"kill `lsof -t -i:%@`", self.port];
+        system([script UTF8String]);
     }
     
 	// build arguments
@@ -308,17 +348,34 @@ BOOL _isServerListening;
     {
 		nodeCommandString = [nodeCommandString stringByAppendingString:@" --no-reset"];
     }
-	if (self.keepArtifacts)
+    if (!self.overrideExistingSessions)
     {
-		nodeCommandString = [nodeCommandString stringByAppendingString:@" --keep-artifacts"];
+        nodeCommandString = [nodeCommandString stringByAppendingString:@" --no-session-override"];
     }
+    if (self.useSeleniumGridConfigFile)
+    {
+        nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--nodeconfig", [self.seleniumGridConfigFile stringByReplacingOccurrencesOfString:@" " withString:@"\\ "]];
+    }
+    
+    // logging preferences
 	if (self.useQuietLogging)
     {
 		nodeCommandString = [nodeCommandString stringByAppendingString:@" --quiet"];
-        
+    }
+    if (self.keepArtifacts)
+    {
+		nodeCommandString = [nodeCommandString stringByAppendingString:@" --keep-artifacts"];
+    }
+    if (self.useLogFile)
+    {
+        nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--log", [self.logFile stringByReplacingOccurrencesOfString:@" " withString:@"\\ "]];
+    }
+    if (self.useLogWebHook)
+    {
+        nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--webhook", self.logWebHook];
     }
     
-    // iOS Prefs
+    // iOS preferences
     if (self.platform == Platform_iOS)
     {
         if (self.useNativeInstrumentsLib)
@@ -353,7 +410,7 @@ BOOL _isServerListening;
 		}
     }
     
-    // Android Prefs
+    // Android preferences
     if (self.platform == Platform_Android)
     {
         if (self.useAndroidPackage)
@@ -384,9 +441,19 @@ BOOL _isServerListening;
 		{
 			nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ \"%d\"", @"--selendroid-port", [self.selendroidPort intValue]];
 		}
+        
+        // Android keystore preferences
+        if (self.useAndroidKeystore)
+        {
+            nodeCommandString = [nodeCommandString stringByAppendingString:@" --use-keystore"];
+            nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--keystore-path", [self.androidKeystorePath stringByReplacingOccurrencesOfString:@" " withString:@"\\ "]];
+            nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--keystore-password", [self.androidKeystorePassword stringByReplacingOccurrencesOfString:@" " withString:@"\\ "]];
+            nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--key-alias", [self.androidKeyAlias stringByReplacingOccurrencesOfString:@" " withString:@"\\ "]];
+            nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ %@", @"--key-password", [self.androidKeyPassword stringByReplacingOccurrencesOfString:@" " withString:@"\\ "]];
+        }
     }
 	
-	// Robot Prefs
+	// Robot preferences
 	if (self.useRobot)
     {
 		nodeCommandString = [nodeCommandString stringByAppendingFormat:@" %@ \"%@\"", @"--robot-address", self.robotAddress];
